third_party/blink/web_tests/wpt_internal/css/css-transitions/parsing/top-layer-valid.html

<!DOCTYPE html>
<title>CSS Transitions: parsing top-layer with valid values</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/8189">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
  // top-layer: none | browser
  test_valid_value("top-layer", "none");
  test_valid_value("top-layer", "browser");
</script>
